date formatting belongs in date util
