+++ src/app/shared/util/date.util.ts
@@ -308,3 +308,47 @@ export const getAmountOfIntervals = (beginDate: Date, endDate: Date, interval: I
       break;
   }
 };
+
+export const UTCDateFormat = (date: Date): string => {
+  return date.toUTCString();
+};
+
+export const dailyDateFormat = (date: Date): string => {
+  const dayOfWeek = (day: number): string => {
+    switch (day) {
+      case 0:
+        return 'Sun';
+      case 1:
+        return 'Mon';
+      case 2:
+        return 'Tue';
+      case 3:
+        return 'Wed';
+      case 4:
+        return 'Thu';
+      case 5:
+        return 'Fri';
+      case 6:
+        return 'Sat';
+    }
+  };
+  return dayOfWeek(date.getUTCDay()) + ' ' + (date.getUTCMonth() + 1) + '/' + date.getUTCDate();
+};
+
+export const dayMonthDateFormat = (date: Date): string => {
+  return date.getUTCMonth() + 1 + '/' + date.getUTCDate();
+};
+
+export const monthDateYearFormat = (date: Date): string => {
+  return date.getUTCMonth() + 1 + '/' + date.getUTCDate() + '/' + date.getUTCFullYear() % 100;
+};
+
+export const hourlyDateFormat = (date: Date): string => {
+  const minutes = date.getUTCMinutes() < 10 ? '0' + date.getUTCMinutes() : date.getUTCMinutes();
+  return (date.getUTCMonth() + 1) + '/' + date.getUTCDate() + ' ' +
+    date.getUTCHours() + ':' + minutes;
+};
+
+export const monthYearFormat = (date: Date): string => {
+  return date.getUTCMonth() + 1 + '/' + date.getUTCFullYear() % 100;
+};
